requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
# pjsua2 must be installed via system packages
# See README.md for installation instructions
